+++ src/app/admin/components/product-list/product-list.component.ts
@@ -1,6 +1,11 @@
 import { Component, OnInit } from '@angular/core';
+
+import { Store } from '@ngrx/store';
+
 import { ProductsService } from '../../../products-feature/services/products/products.service';
 import { ProductItem } from '../../../products-feature/models/product-item.model';
+import { Observable } from 'rxjs/Observable';
+import { Product } from '../../../core/models';
 
 @Component({
   selector: 'app-product-list',
@@ -8,7 +13,7 @@ import { ProductItem } from '../../../products-feature/models/product-item.model
   styleUrls: ['./product-list.component.css']
 })
 export class ProductListComponent implements OnInit {
-  products$: Promise<ProductItem[]>;
+  products$: Store<ReadonlyArray<Product>>;
   constructor(
     private productsService: ProductsService
   ) { }