+++ src/app/core/+store/products/index.ts
@@ -1,0 +1,7 @@
+export * from './products.state';
+export * from './products.actions';
+export * from './products.effects';
+export * from './products.reducer';
+export * from './products.selectors';
+
+