+++ src/app/core/+store/router/router.actions.ts
@@ -1,0 +1,21 @@
+import { Action } from '@ngrx/store';
+
+import { NavigationExtras } from '@angular/router';
+
+export class RouterActionTypes {
+    static readonly GO      = '[Router] GO';
+}
+
+export class Go implements Action {
+    readonly type = RouterActionTypes.GO;
+    constructor(
+        public payload: {
+            path: any[],
+            queryParams?: object,
+            extras?: NavigationExtras
+    }) { }
+}
+
+export type RouterActions
+ = Go;
+