+++ src/app/products-feature/components/product-list/product-list.component.ts
@@ -1,8 +1,15 @@
 import { Component, Output, EventEmitter, Input, OnInit, ChangeDetectionStrategy } from '@angular/core';
 
+import { Observable } from 'rxjs/Observable';
+
+import { Store } from '@ngrx/store';
+
 import { ProductsService } from '../../services/products/products.service';
 import { ProductItem } from '../../models/product-item.model';
 import { Router, ActivatedRoute } from '@angular/router';
+import { Product } from '../../../core/models';
+import { GetProducts, AppState, ProductsState } from '../../../core/+store';
+import * as RouterActions from './../../../core/+store/router/router.actions';
 
 @Component({
   selector: 'app-product-list',
@@ -14,21 +21,23 @@ import { Router, ActivatedRoute } from '@angular/router';
 export class ProductListComponent implements OnInit {
   resolveFn: Function|null = null;
 
-  public products: Promise<ProductItem[]> = new Promise<ProductItem[]>(resolve => this.resolveFn = resolve);
+  public products$: Store<ReadonlyArray<Product>>;
 
   constructor(
     public productService: ProductsService,
     private route: ActivatedRoute,
-    private router: Router
+    private router: Router,
+    private store: Store<AppState>
   ) {  }
 
   ngOnInit(): void {
     this.getProducts();
+
+    this.store.dispatch(new GetProducts());
   }
 
   getProducts() {
-    this.productService.getProducts()
-      .then(resp => this.resolveFn(resp));
+    this.products$ = this.productService.getProducts();
   }
 
   move(product: ProductItem) {
@@ -37,6 +46,11 @@ export class ProductListComponent implements OnInit {
 
   navigate(product: ProductItem) {
     const link = ['details', product.id];
-    this.router.navigate(link, {relativeTo: this.route });
+    // this.router.navigate(link, {relativeTo: this.route });
+    const payload = {
+      path: link,
+      extras: { relativeTo: this.route }
+    };
+    this.store.dispatch(new RouterActions.Go(payload));
   }
 }