+++ src/app/products-feature/guards/index.ts
@@ -1,0 +1,2 @@
+export * from './product-resolve.guard';
+export * from './products-state-preloading';